+++ src/NovelRT.Java/core/src/main/kotlin/com/github/novelrt/ecs/EntityIdVector.kt
@@ -8,8 +8,8 @@ class EntityIdVector internal constructor(
     handle: Long,
     owned: Boolean
 ) : KotlinNativeObject(handle, owned, NovelRT::Nrt_EntityIdVector_Delete) {
-    fun insert(entity: EntityId) = NovelRT.Nrt_EntityIdVector_Insert(handle, entity.toInt()).handleNrtResult()
-    fun remove(entity: EntityId) = NovelRT.Nrt_EntityIdVector_Remove(handle, entity.toInt()).handleNrtResult()
+    fun insert(entity: EntityId) = NovelRT.Nrt_EntityIdVector_Insert(handle, entity.toLong()).handleNrtResult()
+    fun remove(entity: EntityId) = NovelRT.Nrt_EntityIdVector_Remove(handle, entity.toLong()).handleNrtResult()
 
     companion object : SingleTrackingContainer<EntityIdVector>(NativeObjectTracker.Target.UNOWNED_OBJECTS) {
         override fun makeObject(handle: ObjectHandle<EntityIdVector>): EntityIdVector =
